# ProjetMinuteMed/src/main/webapp/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Gives all the details of the MinuteMed APIs",
    "version" : "1.0.0",
    "title" : "MinuteMedAPI"
  },
  "host" : "localhost:8080",
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/auth" : {
      "get" : {
        "tags" : [ "Authentication REST Endpoint" ],
        "summary" : "get all user account",
        "description" : "",
        "operationId" : "findAllUsers",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "201" : {
            "description" : "All the user account."
          },
          "404" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/auth/create/datamanager" : {
      "post" : {
        "tags" : [ "Authentication REST Endpoint" ],
        "summary" : "Create a data manager account",
        "description" : "",
        "operationId" : "createDataManagerAccount",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserAccount"
            }
          },
          "201" : {
            "description" : "The User Account is created."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "409" : {
            "description" : "Username conflict"
          }
        }
      }
    },
    "/auth/create/doctor" : {
      "post" : {
        "tags" : [ "Authentication REST Endpoint" ],
        "summary" : "Create a doctor account",
        "description" : "",
        "operationId" : "createDoctorAccount",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserAccount"
            }
          },
          "201" : {
            "description" : "The User Account is created."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "409" : {
            "description" : "Username conflict"
          }
        }
      }
    },
    "/auth/create/nurse" : {
      "post" : {
        "tags" : [ "Authentication REST Endpoint" ],
        "summary" : "Create a nurse account",
        "description" : "",
        "operationId" : "createNurseAccount",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserAccount"
            }
          },
          "201" : {
            "description" : "The User Account is created."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "409" : {
            "description" : "Username conflict"
          }
        }
      }
    },
    "/auth/delete/{idAccount}" : {
      "delete" : {
        "tags" : [ "Authentication REST Endpoint" ],
        "summary" : "delete an user account",
        "description" : "",
        "operationId" : "deleteUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idAccount",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "201" : {
            "description" : "The User Account is deleted."
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/auth/login" : {
      "post" : {
        "tags" : [ "Authentication REST Endpoint" ],
        "summary" : "login to the application",
        "description" : "",
        "operationId" : "authenticateUser",
        "consumes" : [ "application/x-www-form-urlencoded" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "login",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "201" : {
            "description" : "You are login"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/auth/logout" : {
      "post" : {
        "tags" : [ "Authentication REST Endpoint" ],
        "summary" : "logout to the application",
        "description" : "",
        "operationId" : "logout",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "201" : {
            "description" : "You are logout"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
...
    "/nodes/APHP/hospitals/poles/services" : {
      "get" : {
        "tags" : [ "Arborescence REST Endpoint" ],
        "summary" : "Get all Services of the APHP.",
        "description" : "",
        "operationId" : "getAllServices",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/NodeService"
              }
            }
          },
          "201" : {
            "description" : "All Services are returned."
          },
          "404" : {
            "description" : "APHP not found"
          }
        }
      }
    },
...
    "/nodes/APHP/{idAPHP}/hospitals/{idHospital}/poles/{idPole}/services" : {
      "get" : {
        "tags" : [ "Arborescence REST Endpoint" ],
        "summary" : "Get all Services of a Pole.",
        "description" : "",
        "operationId" : "getServices",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idAPHP",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idHospital",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idPole",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/NodeService"
              }
            }
          },
          "201" : {
            "description" : "All Services are returned."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Services not found"
          }
        }
      },
      "post" : {
        "tags" : [ "Arborescence REST Endpoint" ],
        "summary" : "Create a Service",
        "description" : "",
        "operationId" : "createService",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idAPHP",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idHospital",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idPole",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "201" : {
            "description" : "The Service is created."
          },
          "400" : {
            "description" : "Invalid input."
          },
          "404" : {
            "description" : "Unknow node."
          }
        }
      }
    },
    "/nodes/APHP/{idAPHP}/hospitals/{idHospital}/poles/{idPole}/services/{idService}" : {
      "get" : {
        "tags" : [ "Arborescence REST Endpoint" ],
        "summary" : "Get a specific Service.",
        "description" : "",
        "operationId" : "getService",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idAPHP",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idHospital",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idPole",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idService",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/NodeService"
            }
          },
          "201" : {
            "description" : "The Service is returned."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Service not found"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "NodeAPHP" : {
      "type" : "object",
      "required" : [ "floor", "name" ],
      "properties" : {
        "idNode" : {
          "type" : "integer",
          "format" : "int64"
        },
        "floor" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
...
    "NodeHospital" : {
      "type" : "object",
      "required" : [ "father", "floor", "name" ],
      "properties" : {
        "idNode" : {
          "type" : "integer",
          "format" : "int64"
        },
        "floor" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "father" : {
          "$ref" : "#/definitions/NodeAPHP"
        }
      }
    },
    "NodePole" : {
      "type" : "object",
      "required" : [ "father", "floor", "name" ],
      "properties" : {
        "idNode" : {
          "type" : "integer",
          "format" : "int64"
        },
        "floor" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "father" : {
          "$ref" : "#/definitions/NodeHospital"
        }
      }
    },
    "NodeService" : {
      "type" : "object",
      "required" : [ "father", "floor", "name" ],
      "properties" : {
        "idNode" : {
          "type" : "integer",
          "format" : "int64"
        },
        "floor" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "father" : {
          "$ref" : "#/definitions/NodePole"
        }
      }
    },
...
    "UserAccount" : {
      "type" : "object",
      "required" : [ "password", "type", "username" ],
      "properties" : {
        "idAccount" : {
          "type" : "integer",
          "format" : "int64"
        },
        "username" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      }
    }
  }
}